update view admin
