+++ src/main/resources/templates/artists/list_admin.html
@@ -8,7 +8,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
 
     <!-- Title -->
-    <title th:text="${artist.name} + ' - Artist Detail'">Artist Detail</title>
+    <title>Music App</title>
 
     <!-- Favicon -->
     <link th:href="@{/imgs/core-img/favicon.ico}" rel="icon">
@@ -29,51 +29,57 @@
 </div>
 
 <!-- Header -->
-<header th:insert="fragment :: header"></header>
+<header th:insert="~{fragment :: header}">
+</header>
 
 <!-- ##### Header Area End ##### -->
 
 <!-- ##### Breadcumb Area Start ##### -->
 <section class="breadcumb-area bg-img bg-overlay" th:style="'background-image: url(' + @{/imgs/bg-img/bg-1.jpg} + ');'">
     <div class="bradcumbContent">
-        <h2>Artist Detail</h2>
+        <h2>Artist for Admin</h2>
     </div>
 </section>
 <!-- ##### Breadcumb Area End ##### -->
 
 <!-- ##### Artist Profile Area Start ##### -->
-<section class="artist-profile section-padding-100-0">
+<section class="album-catagory section-padding-100-0">
     <div class="container">
-        <div class="row">
-            <div class="col-12 col-md-4">
-                <!-- Artist Image -->
-                <img th:src="@{'/image/' + ${artist.image}}" alt="Artist Image" class="img-fluid">
-            </div>
-            <div class="col-12 col-md-8">
-                <!-- Artist Details -->
-                <h2 th:text="${artist.name}">Artist Name</h2>
-                <p th:text="${artist.biography}">Artist Biography</p>
-                <!-- Additional Information -->
-                <div class="artist-additional-info">
-                    <p><strong>Profile:</strong> <span th:text="${artist.artistProfile}">Artist Profile</span></p>
-                </div>
+        <div class="row oneMusic-albums">
+            <div class="col-12 col-sm-4 col-md-3 col-lg-2 single-album-item"
+                 th:each="artist : ${artists}">
+                <div class="single-album">
+                    <img th:src="@{'/image/' + ${artist.image}}" th:alt="${artist.name}">
+                    <div class="album-info">
+                        <a th:href="@{/music-app/admin/artists/{id}(id=${artist.id})}">
+                            <h5 th:text="${artist.name}">Artist Name</h5>
+                        </a>
 
-                <!-- Edit and Delete Buttons -->
-                <div class="artist-actions mt-4">
-                    <!-- Edit Button -->
-                    <a th:href="@{/music-app/admin/artists/{id}/edit(id=${artist.id})}">Edit</a>
-
-                    <!-- Delete Button -->
-                    <form th:action="@{/music-app/admin/artists/{id}/delete(id=${artist.id})}" method="post" style="display:inline;">
-                        <input type="hidden" name="_method" value="DELETE"/>
-                        <button type="submit" class="btn btn-danger">Delete</button>
-                    </form>
+                        <!--                        <p th:text="${artist.artistProfile}">Artist Profile</p>-->
+                        <!--                        <p th:text="${artist.biography}">Artist Biography</p>-->
+                    </div>
                 </div>
             </div>
         </div>
     </div>
 </section>
+
 <!-- ##### Artist Profile Area End ##### -->
+
+
+<!-- ##### Album Catagory Area End ##### -->
+
+<!-- ##### Buy Now Area Start ##### -->
+
+<!-- ##### Buy Now Area End ##### -->
+
+<!-- ##### Add Area Start ##### -->
+
+<!-- ##### Song Area End ##### -->
+
+<!-- ##### Contact Area Start ##### -->
+
+<!-- ##### Contact Area End ##### -->
 
 <!-- ##### Footer Area Start ##### -->
 <footer class="footer-area">
@@ -84,6 +90,7 @@
                     <img th:src="@{/imgs/core-img/logo.png}" alt="Logo">
                 </a>
                 <p class="copywrite-text">
+                    <!-- Hiển thị năm hiện tại -->
                     Copyright &copy;<script th:inline="javascript">document.write(new Date().getFullYear());</script>
                     All rights reserved | This template is made with <i class="fa fa-heart-o" aria-hidden="true"></i> by
                     <a href="https://colorlib.com" target="_blank">Colorlib</a>
@@ -117,17 +124,6 @@
 <script th:src="@{/js/plugins/plugins.js}"></script>
 <!-- Active js -->
 <script th:src="@{/js/active.js}"></script>
-
-<!-- Confirmation for delete action -->
-<script>
-    document.querySelectorAll('.btn-danger').forEach(button => {
-        button.addEventListener('click', function(event) {
-            if (!confirm('Are you sure you want to delete this artist?')) {
-                event.preventDefault();
-            }
-        });
-    });
-</script>
 
 </body>
